Combine `OpeningHoursClockIcon` and `OpeningHoursClock` into `OpeningHoursSidebar`
This was the initial idea, and now it works so that there is only one React component that renders both the icon and the large clock.
