--- src/apps/opening-hours-sidebar/OpeningHoursClock.mount.ts
@@ -1,4 +1,0 @@
-import addMount from "../../core/addMount";
-import OpeningHoursClock from "./OpeningHoursClock.entry";
-
-addMount({ appName: "opening-hours-clock", app: OpeningHoursClock });
