--- src/apps/opening-hours-sidebar/OpeningHoursClockIcon.mount.ts
@@ -1,4 +1,0 @@
-import addMount from "../../core/addMount";
-import OpeningHoursClockIcon from "./OpeningHoursClockIcon.entry";
-
-addMount({ appName: "opening-hours-clock-icon", app: OpeningHoursClockIcon });
