+++ src/apps/opening-hours-sidebar/OpeningHoursSidebar.mount.ts
@@ -1,0 +1,4 @@
+import addMount from "../../core/addMount";
+import OpeningHoursSidebar from "./OpeningHoursSidebar.entry";
+
+addMount({ appName: "opening-hours-sidebar", app: OpeningHoursSidebar });
